[1227] Add feature spec to test updating a section and returning to confirm page
