+++ spec/features/trainees/diversities/diversity_progress_spec.rb
@@ -8,38 +8,39 @@ feature "completing the diversity section", type: :feature do
     given_a_trainee_exists(diversity_disclosure: nil, ethnic_group: nil, disability_disclosure: nil)
   end
 
-  scenario "renders a 'not started' status when diversity details are not provided" do
-    and_i_visit_the_review_draft_page
-    then_the_diversity_section_should_be(not_started)
+  context "progress tracking" do
+    scenario "renders a 'not started' status when diversity details are not provided" do
+      and_i_visit_the_review_draft_page
+      then_the_diversity_section_should_be(not_started)
+    end
+
+    scenario "renders an 'in progress' status when diversity information partially provided" do
+      given_valid_diversity_information
+      given_i_visited_the_review_draft_page
+      when_i_visit_the_diversity_confirmation_page
+      and_unconfirm_my_details
+      and_i_visit_the_record_page
+      and_i_visit_the_review_draft_page
+      then_the_diversity_section_should_be(in_progress)
+    end
+
+    scenario "renders a completed status when valid diversity information provided" do
+      given_valid_diversity_information
+      and_i_visit_the_review_draft_page
+      then_the_diversity_section_should_be(completed)
+    end
   end
 
-  scenario "renders an 'in progress' status when diversity information partially provided" do
-    given_valid_diversity_information
-    given_i_visited_the_review_draft_page
-    when_i_visit_the_diversity_confirmation_page
-    and_unconfirm_my_details
-    and_i_visit_the_record_page
-    and_i_visit_the_review_draft_page
-    then_the_diversity_section_should_be(in_progress)
-  end
-
-  scenario "renders a completed status when valid diversity information provided" do
-    given_valid_diversity_information
-    and_i_visit_the_review_draft_page
-    then_the_diversity_section_should_be(completed)
+  context "updating a section" do
+    scenario "returns back to confirm page after successfully updating" do
+      given_valid_diversity_information
+      when_i_visit_the_diversity_confirmation_page
+      and_i_edit_my_ethnic_group_details
+      then_i_am_redirected_to_the_confirm_page
+    end
   end
 
 private
-
-  def given_valid_diversity_information
-    @trainee.diversity_disclosure = Diversities::DIVERSITY_DISCLOSURE_ENUMS[:diversity_disclosed]
-    @trainee.ethnic_group = Diversities::ETHNIC_GROUP_ENUMS[:asian]
-    @trainee.ethnic_background = "some background"
-    @trainee.disability_disclosure = Diversities::DISABILITY_DISCLOSURE_ENUMS[:disabled]
-    @trainee.disabilities = [create(:disability, name: "deaf")]
-    @trainee.progress.diversity = true
-    @trainee.save!
-  end
 
   def and_i_visit_the_review_draft_page
     review_draft_page.load(id: @trainee.slug)
@@ -56,11 +57,19 @@ private
   end
 
   def when_i_visit_the_diversity_confirmation_page
-    diversities_confirm_page.load(id: @trainee.slug, section: "information-disclosed")
+    diversities_confirm_page.load(id: @trainee.slug)
+  end
+
+  def and_i_edit_my_ethnic_group_details
+    diversities_confirm_page.edit_link_for(:ethnicity).click
+    and_i_choose(Diversities::ETHNIC_GROUP_ENUMS[:asian])
+    ethnic_group_page.submit_button.click
+    and_i_choose_a_background(Diversities::BANGLADESHI.downcase)
+    ethnic_background_page.submit_button.click
   end
 
   def then_i_am_redirected_to_the_confirm_page
-    expect(diversities_confirm_page).to be_displayed(id: @trainee.slug, section: "information-disclosed")
+    expect(diversities_confirm_page).to be_displayed(id: @trainee.slug)
   end
 
   def and_unconfirm_my_details
