+++ spec/support/features.rb
@@ -5,6 +5,7 @@ require_relative "features/trainee_steps"
 require_relative "features/dttp_steps"
 require_relative "features/common_steps"
 require_relative "features/page_helpers"
+require_relative "features/diversity_steps"
 require_relative "dfe_sign_in_user_helper"
 
 RSpec.configure do |config|
@@ -19,6 +20,7 @@ RSpec.configure do |config|
   config.include Features::DttpSteps, type: :feature
   config.include Features::CommonSteps, type: :feature
   config.include Features::PageHelpers, type: :feature
+  config.include Features::DiversitySteps, type: :feature
   config.include DfESignInUserHelper, type: :feature
 
   config.around :each do |example|
