+++ spec/support/page_objects/trainees/diversities/confirm_details.rb
@@ -5,6 +5,12 @@ module PageObjects
     module Diversities
       class ConfirmDetails < PageObjects::Trainees::ConfirmDetails
         set_url "/trainees/{id}/diversity/confirm"
+
+        def edit_link_for(section)
+          within(".govuk-summary-list__row.#{section}") do
+            find(".govuk-link")
+          end
+        end
       end
     end
   end
